added tutorials query and categories query
