+++ server/package.json
@@ -5,15 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "node seed.js && jest",
-    "watch": "nodemon server.js",
+    "watch": "node index.js",
     "seed": "node seed.js"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.[t|j]sx?$": "babel-jest"
-    },
-    "globalSetup": "./test/utils/globalSetup.js",
-    "globalTeardown": "./test/utils/globalTeardown.js"
   },
   "keywords": [],
   "author": "",
@@ -25,13 +18,6 @@
     "mongoose": "^6.1.3"
   },
   "devDependencies": {
-    "@babel/register": "^7.16.9",
-    "apollo-boost": "^0.4.9",
-    "babel-jest": "^27.4.6",
-    "cross-fetch": "^3.1.4",
-    "dot-env": "^0.0.1",
-    "graphql": "^15.8.0",
-    "jest": "^27.4.7",
     "nodemon": "^2.0.15"
   }
 }
