--- server/test/Category.test.js
@@ -1,1 +1,0 @@
-const client = require('./utils/getClient');
