--- server/test/utils/getClient.js
@@ -1,8 +1,0 @@
-const ApolloClient = require('apollo-boost');
-
-const client = new ApolloClient({
-  uri: 'http://localhost:4000/',
-  onError: (e) => { console.log(e) },
-});
-
-module.exports = client;
