--- server/test/utils/globalTeardown.js
@@ -1,4 +1,0 @@
-// Stop the development server
-module.exports = async () => {
-    await global.httpServer.stop();
-};
